deployment/requirements.txt
Flask==2.3.2
pandas==2.0.3
dash==2.11.1
dash-bootstrap-components==1.4.1
plotly==5.15.0
gunicorn==20.1.0
redis 
fastapi-cache2
celery